warnings
warnings (add command helpers)

+++ src/bklib/spatial_map.hpp
@@ -99,7 +99,7 @@ public:
     void remove_if(Predicate&& predicate, Remove&& do_remove) {
         for (auto i = 0u; i < sorted_.size(); ) {
             auto const& p = sorted_[i];
-            auto& d = data_[p.second];
+            auto& d = data_[static_cast<size_t>(p.second)];
             if (predicate(p.first, d)) {
                 auto const size = data_.size();
                 do_remove(p.first, d);
